Change RAID level to u16.

+++ src/sim_engine.rs
@@ -17,7 +17,7 @@ impl Engine for SimEngine {
     fn create_pool(&self,
                    name: &str,
                    blockdev_paths: &[&str],
-                   raid_level: i32)
+                   raid_level: u16)
                    -> StratisResult<Box<Pool>> {
         println!("sim: pool created");
